Limits duration waiting for transactions to close when stopping db
Stopping a db will have it wait for any still closing transactions
to be fully closed. Previously this time was unbounded and could
conceivably block/hang on e.g. acquiring locks or calling
transaction event listeners. Now there's a max wait time of
what the setting 'dbms.shutdown_transaction_end_timeout' before
continuing the stopping of the database.

Functionally the end state of the database in a scenario where there
are transactions still closing when performing the final checkpoint
is similar to running a "live" checkpoint on a running db and
crashing soon thereafter.

The guarantee of a clean shutdown is sacrificed for the ability
to stop a database that has indefinitely hanging/blocking closing
transactions.