Corrects some issues in ReuseStorageSpaceIT
It would create no data at all if number of available processors
on the machine running it was >= 10.

Also only compares ID generators that handles single IDs,
because there's no guarantee of perfect reuse for ID generators
that manage ranges of IDs.

Also removed IdType#isSchemaType() which was introduced temporarily
to accommodate for a test, but no longer seems to be used.