FulltextIndexProviderTest fixes to run Freki
This includes a Freki bug where wrong index updates would be generated
whenever new indexed relationships would be added, one or more
lower existing types were sparse and the new type is sparse.
The changes pointing out which ordinal relationships were to be indexed
in IndexData inside the Relationship command would get the wrong
ordinals due to a previous optimization when copying a whole
sparse group that can no longer be done to the same extent due to
the need to count the number of copied relationships.