Fix SlottedTrailPipe sometimes returns results in wrong columns ()
* Fix a case where SlottedTrailPipe would return results in the wrong column.

When mapping over the set of slots to create getters and setters and then calling toArray, sometimes when there are multiple groupNodes the order of the two resulting arrays would be different. The solution is to create an array first and then map over that array twice. The actual order of the array is not important, but preserving the order of the array is.

* Skip intermediate private array with sorting and just change the input params to Array instead