Fixing a flaky fulltext test
The test used to check that the result entities from the index come
in a particular order. The test assumed that an entity created first
would have a lower ID than an an entiry created after it. This is not
true in this test, because its test class reuses one database for multiple
tests, so no assuptions can be made about the reused entity ID allocation.