Optimise degree(n) > 0
This is a weird and non-intuitive optimisation. For some reason it's faster to
calculate degree(n) > 0 in dense nodes using a cursor and calling next(),
instead of asking kernel for the degree. For sparse nodes it's faster the
other way around.

This optimisation makes it possible for planner to start plan queries like
`match (a:A) where exists((a)-->())` using a filter instead of
`SemiApply` + `Expand`.

Co-Authored-By: Satia Herfert <satiaherfert@gmx.de>