Fixes issue with per-group id type for full import
Incremental import has added the ability to specify the type of
ID values. This also affected full import, although full import
doesn't have an as flexible ID mapping and so would fail on different
types of id values.

This commit fixes that problem so that the command-line-provided
`--id=type` is used for the in-memory ID mapping for full import,
but the type of the ID values, if stored on the nodes, can be
specified using the `{id-type:<type>}` construct in the CSV header.