Introduce chunked transaction commit process ()
* Introduce chunked transaction commit process

Before this change set transactions were always completely accumulated in memory before
applying them to the database. In addition to the good things that come with problems like OOM, non-ability
in principle to execute transactions bigger than available memory, etc.

With snapshotting possibilities of new multi-versioned stores we now have abilities to actually start writing
transactions while they are still running in the database without compromising data consistency for ongoing transactions.
With those features available for us, we introduce here a new commit process for such stores that are able to write transactions in chunks.

To be able to do that some changes are required to the available code:
 - the ability to append some set of commands, not only full transactions;
 - detach transaction id generation from the commit-append process;
 - detach transaction commitment from the commit-append process;
 - adapt transaction code to be able to select proper commit possibilities;
 - adapt transaction state to be able to chunk;
 - perform some invasive renaming to adapt to new concepts.

The new process for chunked transactions will split transactions into a number of chunks.
The chunk is ready as soon as the transaction state reaches some specified memory boundary (setting controlled).
After that transaction state is converted to commands and applied to the store but without actually committing that transaction.
And only when the user will initiate a commit transaction will actually be committed and visible to other transactions.
Chunked transactions will have its transaction id generated as part of the first chunk generation. As a result that will have
an influence on the transaction committed sequence and as well recovery, since chunks may be interleaved.
But none of those 2two problems are solved so far in this PR. This PR only adds support for single-writer transactions
that do not know how to recover from failure or rollback. Those features will be added separately.

* Cleanup mock leftovers for CommittedTransactionRepresentation.

* Restore validation of ids in default id generator and restore tests

* Pass correct id generator in tests

* Review comments:
- use suggest way of detecting multi versioned db;
- rename method

* Review comment: missing `o`

* Adapt mock based tests

* Fix dead freki test: incorrect id generation during recovery for count store applier