Rework of ConcurrentAccessIT to hopefully make it faster
ConcurrentAccessIT takes ~10min in CI. It was using a CyclicBarrier
with calls to its await() method in various places. Not quite sure
it was used properly, so simplified it using a DoubleLatch
as well as lowered num workers/requests. It still tests the same thing,
which is to make sure that bolt can serve concurrent connections.