Eagerness analysis for quantified path patterns
Handle quantified path patterns as part of eagerness analysis.
This approach of adding the query graph of the quantified path pattern to the list of all query graphs is probably overly conservative and can be refined.
Note that this commit only addresses the intermediate representation eagerness analysis, leaving out the logical plan one.