refactor: reorder parameters of KernelVersion plumbing
Original commit was plumbing KernelVersion into KernelTransaction
instances. 939c6c9a939a1d75283a3ff3a31cb4fb0b77adbd also did similar
plumbing; therefore after rebasing this mainly reorders parameters in a
way they will make more sense in future planned refactors in this area.

When transaction is reused, initialise with current version.

KernelTransaction instances will need to be aware of the KernelVersion
they correspond to. This is such that they can request commands in that
version. Will aim to store within CommandCreationContext.

As commands are only created during commit, we could instead plumb the
KernelVersion.Provider into the KernelTransaction, and only retrieve the
version just when needed. That idea was scrapped, when learning that
there is a plan to chunk tx state and commit early (before commit), at
least in MVCC; therefore, the KernelVersion needs to be known when the
KernelTransaction is initialised.