Block migration of pre-5 enterprise system db
Also fixes so that migrate '*' actually upgrades the system database (when it should).

About the reason for blocking migration:
There is a SystemGraphComponent that doesn't have an upgrade path from all versions.
It was decided a while back that this was not a supported migration path. This just
blocks so users don't try to use it accidentally.