Freki relationship creation doesn't read X1 for added nodes
For this the RelationshipIdGenerator needs to know if the nodes
are added in tx state and skip the read if so.

In Operations#relationshipCreate and the calls it made there were
already at least two calls to
ReadableTransactionState#nodeIsAddedInThisTx (which is a HashMap lookup)
per node and so it was already worth doing those lookups once and
passing in, instead of passing in tx state into each of those methods
for them to look that up individually. Adding a third would have been
a bit over the top.

Also pulling out that state earlier allows for skipping unnecessary
node existence validation for nodes added in the tx.

So the benefit is two-fold:
- Freki gets cheaper relationship creation for nodes that are
  also created in the same tx
- Regardless of storage engine, less "is this node added in this tx"
  lookups are made when creating relationships