Discard unused columns in projections
Adds `discardSymbols` to the projection plan to indicate
variables that can be discarded.
This information is provided to enable runtime optimisations
and improved memory management.
WITH clauses are now planned with these discarded variables.

For example
`... COLLECT (n.value) AS reallyBig WITH size(reallyBig) AS size ...`
will discard the `reallyBig` and enable future runtime changes
to free that memory.