Make and effort to remove holes in ResourceManager
When you close things in the wrong order there may appear holes in the resource tracker.
For long-running queries which accumulate a lot of resources this can lead to the underlying
array having to grow unnecessarily large eventually leading to a MemoryLimitExceededException.
This PR adds a small effort to remove holes that appear when closing resources in the wrong order.