Fix bug where round(x, 0, 'HALF_UP') returned wrong result.
If the user has not requested a rounding mode and the precision is 0, we still want to fallback to Java Math.round(), but that does not have the rounding mode 'HALF_UP', but a custom one where it round ties towards positive infinity.

If the user has request a precision >= 1 and no mode, we still want to use 'HALF_UP' to not break current behavior.