SHOW TRANSACTIONS should show information about internal transactions
We closed the transactional contexts before committing the transaction. When
the context is closed we also unbind the executing query which is the only thing
connecting the inner transaction to the running query and the outer transaction.
The effect is that when running SHOW TRANSACTIONS the internal transaction will be
empty (no "currentQuery" and no "outerTransactionalId") throughout the commit phase,
which in many cases is the majority of time spent.

There is still a brief period of time between when we create the inner transaction and
when we bind it to the executing query during which the transaction will still appear as
empty.