Avoid calling checkpoints on an already shutdown database. ()
Previously it was possible to request checkpoints on an already shutdown database.
Depending on your position in a race you can observe different things. Most probably outcome was some sort of exception.
This PR changes things in a way that on shutdown checkpointer will be marked as "shutdown". That information will be used to prevent any subsequent checkpoints.