Fixing issues with TX log pruning
That were a couple of issues. Some of them usually cancelled each other.
That might be the reason why they were not detected for a while.

The common issue was range inclusivity or exclusivity.
As stated in the javadoc, Threshold.reached should return true for a log
file "below which it should be safe to prune log files". Three implementations
worked like that, but not EntryCountThreshold, which returned true for the first
file that should be deleted.
Another range inclusivity issue was that the result of Threshold.reached
was used to construct INCLUSIVE range of to-be-deleted files.
Another issue was that the last log file was not submitted to Threshold.reached
for evaluation. It means that its size, TX timestapmps or TX count did not participate
in the threshold calculation.
Lastly we kept at least two files:
"LongRange.range(lowestLogVersion, min(thresholdResult.logVersion(), upToVersion - 2))".
This had historic reasons and is not necessary with detached checkpoints as 'upToVersion'
always points to a non-empty TX log file now.