Recovery#performRecovery unregisters config listeners
The main case for Recovery#performRecovery is in Database#start
where the database's config is used. However there are a couple
of other cases in clustering where performRecovery is called
with the global config instance. In these cases some
components participating in recovery would register config listeners
and then not unregister them, making them accumulate more and more
over time as topological changes occurred in the cluster.

This fix is about having Recovery wrap the provided config
in a LocalConfig and removing all listeners when done.
There could be made a point about discipline for each component
to unregister its listeners, but not all such components are
lifecycle components and is hard to keep track of going forward
as well.