Unflake LegacyClusterFactoryTest
By changing how TestDirectory works for nested classes
i.e. the @Nested annotation). Previously the nested test class would
get a new directory and push that to a stack of directories
managed by a single TestDirectory. This created a weird scenario
for e.g. nested cluster test classes (where a cluster extension is
declared in the parent class) would start clusters in the
nested directory, but after all tests in the nested class
then no clusters would be shut down (due to junit extension value store
checking parent on "get", but not on "remove" stored value"",
but the directory would be deleted - which would ultimately fail
on Windows (at least on Windows Server 2019).

Now TestDirectory has only a single directory and does ref counting
on additional prepare/complete so that everything works with the
parent directory, which is probably what is generally desired.