5 prepare raft module ()
Remove dependency on dbms module in Raft components. This is to prepare our Raft code to be moved to its own module.

* Create interface for RaftMemberAddressResolver

The reason for this is so that RaftOutbound does not have a depency on discovery

* Move PortConnectivityChecker to cluster common

* Change name of PrimaryCountProvider and move it to a raft package

* Remove ReadOnlyDatabaseChecker from RaftGroups

It is replaced by a BooleanSupplier indicating that replication is restricted.

* Move panicker to common module

The panicker api will be used by modules that does not want to depend on the dbms module.