Improves merge concurrency for lucene indexes population
As of recent Lucene index population uses a "serial" merge scheduler
in order to avoid over-using CPU resources w/ background merge threads.
However, this merge scheduler is too serial. In index population
all the workers can add data to the Lucene index, but most of the
time was spent waiting for other threads to do their serial merges.

This commit introduces a merge scheduler which performs merges
on the worker threads, but can run merges concurrently with other
worker threads merges. This basically restores the performance back
to that of the concurrent merge scheduler w/o the downside of
spawning additional background threads.