Raft binder break up ()
Sepearates RaftBinder and RaftBootstrapper. Further more it completely rewrites the RaftBinder internals to follow the logic of a state-machine.

RaftBinder is now an interface. The reason for this is that Raft should eventually be moved to its own module and how to bind is something that depends on external components. However, starting the Raft group which is done by RaftBootstrapper should still be within the raft module.

TopologyServiceRaftBinder is the current only implementation of the RaftBinder. It has been rewritten internally so that it transitions between states instead of traverse a tree of methods like RaftBinder did before. This is to provide a logical structure between snapshot updates and state transitions. It should also make the process easier to read and modify in the future.

BoundState will now have one of three outcome which are made clear by its updated API:

IsRestating: This means that this RaftMember had previous raft state and is just resuming where it was. No bootstrap or snapshot download required.

ShouldBootstrap: This means that this member should bootstrap raft (this was previously done inside RaftBinder). This always has a none empty list of initial members.

NeedSnapshot: This means that this member should join a fresh RaftGroup.

Note: Previously BoundState made no difference on NeedSnapshot or IsRestarting, this worked because SnapshotService.awaitState() would never enter the while loop. I changed this because I don't think this should rely on some implementation of a method in separate class so we now skip await state entirely if this is the case.

The new RaftBinder will publish RaftGroupId as soon as a quorum is discovered then it will move over to a Completable state and wait for the last member. This means that if the third member (which has the wrong seed) discovers the published RaftGroup before seed validation then it will join as a none boostrapping member.