fix: required kernel version for transaction
This commit is indended to be _reverted_ in the future, or another
approach taken.

Ideally the transaction is aware of its KernelVersion before it starts
creating commands. A previous refactor in this series acomplished this
by setting the version on the initialisation of the transaction. This
was important, as the MVCC work could potentially write out commands at
any point during the transaction.

Unfortately this would break the current implementation of the rolling
upgrade. By fixing the kernel version at the start of a transaction, the
first write transaction after upgrade will not have access to the new
features. It would be this transaction that would create the upgrade
command.

RelationshipUniquenessConstraintOnDatabaseUpgradeTransactionIT::createRelationshipUniquenessConstraintShouldTriggerUpgrade
when 'enabled' will detect this as an issue that the transaction is not
the right version to use the feature.

We will need to investigate how best to solve this.

It is important that after upgrade, only the new version commands are
being used; otherwise newer functionality or stores will not be
correctly in sync.

Likewise, it should be important that all commands of a transaction are
the same version for the same reason.

For now, the KernelVersion has been switched out to the
KernelVersion.Provider to keep things working as before.