Add internal setting for duplication of user log messages to debug log ()
The default value is true (i.e. duplication enabled).

In most places where tests instantiate a SimpleLogService,
duplication is set to disabled. A handful of tests fail when doing this,
because they assert on the debug log to find messages that are
actually logged to the user log. While some were easy to fix,
the log service in TestDatabaseManagementServiceFactory keeps
duplication enabled.