Migrate store command always migrates to latest kernel version ()
Even if the store is on latest format it might not be on the latest kernel version.
It is very weird for someone that have used the migrate command to possibly need
to call dbms.upgrade procedure to get the latest features.

This makes migrate command always upgrade to latest kernel version.
It takes some shortcuts to not have to go through the whole migration logic
and create dirs/move around files. These shortcuts work for now but at some
point it can be moved into the regular migration flow and just make the migrators
handle no store changes.