Drop test database from `@EnterpriseDbmsExtension` after a successful test ()
Before this PR databases after successful tests were only stopped. Now when we can have a database that actually produced a lot of transaction logs that can cause problems on build agents, and they will run out of disk space.
This PR changes that behavior and drops the database on a successful test, while still only stopping it on failure to keep the stores and state.