Freki importer builds relationship type lookup index separately
which is much faster than building it in-place.
There's a fast version that just reads marker bits of nodes
and which relationship types they have (if there are few types
in the import). And then a "slow" version which builds it
from reading the store. Even the slow version is typically
much faster than building it in-place during the relationship import
and the fast version is a lot faster.