Less GBPTreeWriter overhead for parallel writers
By doing a couple of things for parallel writers:

- To some extent do what the "batched single threaded" writer does,
  which is to only move the cursor the minimum amount between writes.
  This prevents the overhead of starting from the root for every write.
  To not prevent other writers from doing pessimistic writes
  then every e.g. 10th write still resets up to the root.
- Avoid churning latches a bit by having each writer keep references to
  their latches, one per level, so that the part of the tree that the
  writer goes down that is the same as the last time will avoid
  looking up/creating latch instance (which is often the far greater
  overhead) compared to acquiring the latch.

As an example comparing inserting sorted data with the "batched" writer
vs single parallel writer would be ~12X. This overhead is now down
to 10% only. (This is w/o any page faulting).

Tho the overhead still grows the more random/further apart writes are
for a writer, but in that case the overhead also grows
(although not as much) for the "batched single writer".

Note that even for completely random writes the root will most often
still be the same between operations and so at the very worst case
one level worth of overhead is gone.