Use ASTRewriter postConditions in AstRewriting
noDuplicatesInReturnItems was unused expect here, and is actually a SemanticCheck and in no
way connected to ASTRewriter.

noReferenceEqualityAmongVariables is broken in later phases, e.g. the Namespacer. That is not easy to fix.
It seems to work, though. In a separate investigation, we should find out when we actually need
no reference equality among variables. I could remove _all_ cases of `copyVariables` and
_most_ cases of `variable.copyId` with no test failures.