Lift classloader out of ProcedureCompilation ()
.. to make it possible for a consumers of the class to specify which
classloader to use for the procedure compilation. Instead, we let the
ProcedureCompiler inject this value since it can be expected to have the
knowledge.

In this commit, we change the class loader hierarchy for plugins, so
that the compiled procedure/function classes use the URLClassLoader as
its parent, rather than the system class loader. This should provide a
path for garbage collection of procedures once they are no longer in
use.

For builtin procedures we keep the current behaviour of using the system
class loader, since we will not expect to garbage collect these.

Finally, we refactor ProcedureJarLoaderTest to remove some duplicate code.