Fix issue with console appenders in server-logs.xml
When running in daemon mode, `neo4j start`, we remove all console appenders since there is no platform independent definition of output for daemon processes.
When running in console mode, `neo4j console`, this should not happen.

This PR fixes a problem where the console mode is not propagated properly down to the logging framework.

Closes #13030