Fix calculation of start position in log entry reader ()
* Fix calculation of start position in log entry reader
* Rename confusing method name
* make delegate call markAndGet if available

The log position is marked before the kernel version is read. In the case where a file rollover occurs, this will be marked at the end of the channel rather than at the start of the new channel. For this fix, if the log version changes, the position is re-marked (minus 1-byte) to get the correct value

---------

Co-authored-by: Rory Steele <rory.steele@neo4j.com>