fix: neo4j-admin database info --from-path
When --from-path was being used, it could fail during the path
validation. It would attempt to check if the given path was not a
directory of store files; however, it didn't do this correctly.

It would extract the directory name from the path; and treat that as a
database name with the normal Neo4jLayout. So:
    --from-path=/somewhere/completely/different

would effectively check (if default config):
    #NEO4J_HOME/data/databases/different

Clearly, this is not indended.

This changes the semantics such that if --from-path is being used it
overrides the config to point the databases directory to the given path.
The rest of the command then uses the layout for the paths it needs.

The validity check specifically creates a layout that will interpet the
given path as the path to the store files.