Fix flakyness and improve tests
Some BFS tests are flaky on freki and in the past we have rewritten to use
`containsExactlyInAnyOrder` but that means we are not testing the BFS part of
the cursor. This PR rewrites test to handle both that at a given depth the nodes may
come in any order while still asserting that cursor is indeed doing breadth-first search.