[1wUXB3gt] Revert setting name back to dbms.max_databases from server.max_databases ()
The setting was in fact used in a DBMS-wide way at database creation time. For that purpose setting it differently on different servers would give different experiences depending on which one was system leader.

The setting was also used in a per-server way when _starting_ a database, in DatabaseManager, but that check is a leftover from 4.4 behaviour before we had explicit allocation, so it has been removed.