Fix a race in EncodingIdMapper - trackerCache could get corrupted
There was a race in the collision detection where two threads could
be swapping against the same element of the trackerCache (in DetectWorker.detect)
when checking collisions over the seams. The result would be an element missing from
the cache and another being there twice.
Only in the unlikely scenario that two detect works where swapping against the same element
at the same time, so if there were conflicts over the seam and there data index were in the wrong order.