After commit post-gap transaction visibility  ()
* After commit post-gap transaction visibility

Before this PR any transaction that was committed after the gap in the transaction sequence was not visible until the gap will be closed and all missing transactions were completed. That behavior was at least confusing and was not really even on read committed level.
This PR changes the visibility barrier check and allows such transactions to be visible. To be able to do that snapshot now captures not only the highest gap-free transaction id but also a set of closed post-gap transactions that a closed at the moment of transaction initialization.

* Start rework to contain the highest visible and range of not completed transactions instead

* Start changing VersionContext to use the highest visible and not yet visible array instead.

* Start re-adapting conditions to new boundaries (except trees)

* Lets conditions of visibility a bit

* Javadocs update

* Simplified page load conditions

* Review comments: use binary search, rename to reverse snapshot

* Review comments: move cursor methods, some additional renamings