Move AmbiguousAggregationAnalysis after SemanticAnalysis.
This fixes queries like `MATCH (a:A) RETURN count(a) ORDER BY EXISTS { (n:N) }`.
At the same times, having moved the phase also after PreparatoryRewriting,
suddenly allows queries like `MATCH (n) RETURN n.num + 2 AS num, count(*) AS cnt ORDER BY n.num + 2`,
which were previously forbidden.

This makes it possible to completely remove `OrderBy.checkAmbiguousOrdering`.
Either an expression (or a sub-expression) in ORDER BY exactly matches a grouping key from the WITH/RETURN, or we will get the
"In a WITH/RETURN with DISTINCT or an aggregation, it is not possible to access variables declared before the WITH/RETURN" error instead.