Cache StandardAccessMode ()
StandardAccessMode is fairly expensive to construct and this occurs for every new transaction. This introduces caching for it on the same basis as we cache roles from the disk - they will be cached until a write to the system graph occurs.

A StandardAccessMode is not always cachable however as it hold references to node, relationship and property IDs from the Token store. If for some reason when it is constructed the token store is unavailable (for DBMS commands etc.) or of there is no token for a particular label / rel type / property due to none having been created yet, we cannot cache the StandardAccessMode.