Fabric planning uses the same set of procedures during planning
.. to make it possible to use ProcedureViews to represent the
point-in-time available procedures for the entire planning step.

This commit makes a FabricTransaction take a ProcedureView upon creation
and expose it for its planner with an added `transactionlessProcedures`
method.

The planner uses the ProcedureView in two steps.

1. The SignatureResolver uses the procedure view to look up procedure
   tokens, in order to parse the query. Bolt uses e.g. `CALL
   dbms.getRoutingTable()` as a part of its keep-alive.

2. The StaticEvaluator/UseEvaluator use the procedure view to evaluate
   functions for the `USE` statement.

When evaluating a Fabric query there is still a chance of an impedence
mismatch between the preplanning, and the actual evaluation.

The cause of this is that Fabric will dispatch the query along to a
local, virtual, or remote cypher planner to execute each fragment. In
all three cases, a new underlying transaction is formed when actually
executing the fragment, meaning that it is possible to get a new
point-in-time ProcedureView.