Removes importer configuration confusion regarding page cache
Configuration has pageCacheMemory() and maxOffHeapMemory().
pageCacheMemory()'s purpose is to be able to limit the page cache
for the record storage engine below the upper limit of 1GiB,
whereas maxOffHeapMemory() controls the amount of additional off-heap
memory the importer could use.

And since pageCacheMemory() is so specific to the record storage
case the Freki importer uses maxOffHeapMemory() to control its
page cache size (page cache is essentially off-heap, right?).

Confusion sets in for tools and how to use those two.

This commit moves pageCacheMemory() into the record storage importer
and so now there's only maxOffHeapMemory() to configure.