Handle FOREACH conflicts in IR Eager analysis by unwrapping FOREACH
Previously we were very conservative when it comes to FOREACH,
but were still missing some cases.
To not miss these cases, one option would have been to remove
`qgWithInfo.hasUnstableLeaves &&` from `foreachOverlap` but
that lead to many unnecessary Eagers.

So instead, this appraoch flattens updates in FOREACH before
Eagerness analysis, which catches the missed cases, and does
not lead to so many unnecessary Eagers.