Avoid random file access when reading zip-compressed CSV files
Reading CSV data from a ZipFile, first the "correct" entry
is located and then a stream is provided for that entry.
The problem is that some file systems, e.g. streaming from S3 bucket
cannot handle this random access.

However, if doing this using only ZipInputStream then compression ratio
and entry size cannot be determined, something that destroys estimates
and progress (in import).

Therefor this has now been changed to opening the ZipFile to
discover the correct entry, close the ZipFile and then opening
a ZipInputStream, skipping to the known entry - avoiding going
back and forth in the file.