LogEntryParserSets throws exception sympathetic to log reader
When VersionAwareLogEntryReader reads a log entry and it comes
across a versionCode that it doesn't recognize the
KernelVersion.getForVersion() call will throw
IllegalArgumentException which is then caught and converted
into an UnsupportedLogVersionException (which is handled well outside).

However, if the versionCode happens to be 127 (that of the
intermediary-glorious-future version) then instead the
LogEntryParserSets.parserSet() call will throw IllegalStateException
and therefore not be handled properly at all - in the end resulting
in recovery not recovering any transactions att all, even if it
was only failing on the last random bytes (i.e. 1/255 chance).

Made LogEntryParserSets.parserSet() throw IllegalArgumentException,
just like KernelVersion.getForVersion() so that error handling
is the same for both.