Update tree visibility according to the new rules ()
This PR aligns visibility condition when searching into the tree value history to the condition of the store pages.
Value is visible if it's created on or before highest closed and not by not-visible transaction.

Bit-merging version of TreeValueHistory is removed because it doesn't work with new rules due to patching versions that can be considered visible.
Instead users of the tree should take possible conflicts into account and avoid them.

Example for token indexes: if one key-value is responsible for the nodes that sit in the two different pages (a) and (b) it can happen that changing token index for nodes sitting in a page (a) can result in loss of information in the token index for the nodes from page (b). Pages (a) and (b) are consecutive in this case.
Assuming that we resolve conflicts on store page basis (one page version contains only changes from the single transaction), token index only need to ensure that any two nodes that are encoded by the same token index key always belong to the same page.