DbmsRuntimeVersionTest runs with slightly later versions
This test was previously succeeding by accident where the database
on the very old version actually failed to start, a silent error
because the db wasn't retrieved in the test. This was noticed
because Freki would not fail on startup
(due to lacking kernel version checks) and would instead fail
writing a checkpoint of that very old version.

Now it works on both storages.