Initialise KernelTransactionImplementation.initializationTrace to NONE
KernelTeransactionHandle.transactionInitialisationTrace() does not look like it should ever return null. If a new KernelTransactionImplementation is pooled but not yet initialised() it seems like it can do this. When the tx is reset, it is set to NONE so this seems like the correct initial state also.

This comes from seeing in intermittent errors in DestructiveQueriesBoltTest