Freki makes use of all short-string codecs
A long time ago UPPERHEX/LOWERHEX short-string codecs were introduced,
but were added in a way where they would never be selected.
Reason was the order in the enum in which they appeared, i.e.
after (larger bit-wise unfortunately) encodings that could also
encode whatever they could encode.

To not interfere with record-storage format these additional
codecs are now accessible, but only if a flag is passed in as true
when encoding, and Freki does just that.