TransactionLogsRecovery doesn't hold onto command batches unnecessarily
TransactionLogsRecovery had fields for last transaction and
last reversed transaction that it had seen. When recovering very large
transactions it could mean that at the time of recovering one of
those very large transactions, another two very large transactions
could be retained at the same time, preventing them from being
garbage collected and potentially running out of memory.

This could result in a situation where there was enough memory to
to create and commit a transaction, but an insufficient amount of
memory later when recovering it (after a crash or similar).