Keeps root latch referenced for performance
The majority of overhead of the latches is to atomically create them
in the concurrent map. All (parallel) write operations acquires
the root latch so there's a big churn on that latch.
Letting the RootLayer reference that latch keeps it in the latch map
and avoids it being churned. For highly concurrent small write batches
this makes a big difference in performance.