5 kernel versions ()
Throw exception whenever recovery is attempted without transaction logs for primaries and secondaries. There must be transaction logs to guarantee that the correct kernel version is picked. 

Only the restore command is allowed to run without any transaction logs. If forced recovery is enabled, the latest kernel version in the config will be used.