Backport: Checkpoint before listing store files during online backup
This resolves a bug where some Lucene index data would be lost during full online backup.
The problem was that the Lucene index snapshot that is taken when listing unrecoverableFiles was taken before the checkpoint. Any concurrent updates that happen after taking the snapshot but before the checkpoint would be missed by that backup.

  Snapshot    Checkpoint
----|-----------|----->
    ^-----------^
    Missed updated

Add a new test for testing online backup with various indexes.