Fix race in LazyProcedureView initialization ()
.. which could happen when two calls to `LazyProcedureView.initView` would
occur simultaneously.

The reason for the race is that an initialization in one thread might
not have visible for other threads, causing a new initialization to
occur. If a new `ProcedureView` would be created in the intermittent
period, we might end up with two different views being used.

Also renamed the `LAZY_PROCEDURE_VIEW` to `lazyProcedureView`
to follow naming conventions for non-static constants.