Fix NullPointerException in IndexPopulationJob.stop
There is a small race where IndexPopulationJob's jobHandle hasn't been set yet and the
job gets to run and is stopped. It is a benign race since we will wait for the
IndexPopulationJob:s to get to done in stop() of IndexPopulationJobController, which
means there is not really a need to cancel the handle. The population jobs will do
as little as possible when seeing the stopped status.

The found race was benign but resulted in a NPE. This fixes the NPE and moves
around some handling of stopped to get the jobs to abort both faster and
always when stop has been called.